Move TopK shape inference into a reduce submodule
Move the TopK operator, plus its test, from ops.rs into a new
ops/reduce.rs submodule, matching how this operator is grouped in
src/ops.rs.
